+++ packages/graphql-client/tools/update-graphql-schema.js
@@ -4,7 +4,7 @@ const path = require('path');
 const { request } = require('graphql-request');
 const { introspectionQuery } = require('graphql');
 
-const httpEndpoint = () => 'http://localhost:8211/api';
+const httpEndpoint = () => 'http://localhost:8210/api';
 // const httpEndpoint = () => 'https://test.abtnetwork.io/api';
 
 (async () => {